Corrections

+++ src/main/resources/db/migration/V1.1__create_employee_table.sql
@@ -1,6 +1,6 @@
 
 CREATE TABLE `Employee` (
-	`Employee_ID` int auto_increment NOT NULL PRIMARY KEY,
+	`Employee_ID` smallint auto_increment NOT NULL PRIMARY KEY,
     `Name` varchar(100) NOT NULL,
     `Salary` decimal(11,2) NOT NULL,
     `BankNo` varchar(9) NOT NULL,
