+++ src/main/resources/db/migration/V2__create_client.sql
@@ -1,6 +1,6 @@
 create table `Client` (
-`Client_ID` int auto_increment not null primary key,
-`Employee_ID` int not null,
+`Client_ID` smallint auto_increment not null primary key,
+`Employee_ID` smallint not null,
 `Name` varchar(100) not null,
 `Address` varchar(100) not null,
 `PhoneNumber` varchar(11) not null
